refactor(db,replica): extract sync/checkpoint/monitor subsystem (#1221)
Implement Phases 1, 3, 4, and 6 of the incremental extraction proposed
in issue #1221 to reduce structural coupling and prevent cascading fixes.

Phase 1 - SyncResult: verifyAndSync() and sync() return a syncResult
struct instead of mutating db fields as side effects. State is applied
explicitly via applySyncResult().

Phase 3 - Checkpoint phases: Extract checkpoint() into checkpointPreSync,
checkpointExec, checkpointPostSync. Always snapshot in post-checkpoint
sync to capture pages moved by the application's auto-checkpoint during
the db.rtx release window.

Phase 4 - ReplicaScheduler: Extract syncScheduler from monitor().

Phase 6 - Restore: Move ~1200 lines to restore.go.

Fix: Downgrade TRUNCATE to PASSIVE when pre-sync can't reach WAL end.
Fix: Force snapshot on PrevFrameMismatch (WAL changed between verify/sync).
Fix: Add snapshot page count validation in writeLTXFromDB().
Fix: Clear syncedToWALEnd before execCheckpoint releases db.rtx.
